Use more significative test checking package and inner class
